pyproject.toml
[project]
name = "momoa"
version = "0.1.2"
description = "A library for definition, validation and serialisation of Python objects based on JSONSchema specifications."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Berislav Lopac", email = "berislav@lopac.net"},
]
requires-python = ">=3.8,<4.0"
dependencies = [
    "statham-schema>=0.13.5",
    "stringcase>=1.2.0",
    "python-dateutil>=2.8.2",
]

[tool.pdm.build]
includes = ["momoa/"]

[tool.pdm.dev-dependencies]
docs = [
    "mkdocs>=1.4.2",
    "mkdocs-material>=9.0.9",
    "mkapi>=1.0.14",
    "jinja2>=3.1.2",
]
test = [
    "pytest>=7.2.1",
    "pytest-cov>=4.0.0",
    "pytest-spec>=3.2.0",
    "mypy>=0.991",
    "mutmut>=2.4.3",
    "tox>=3.28.0",
    "tox-pdm>=0.6.1",
    "hypothesis-jsonschema>=0.22.0",
]
checks = [
    "black>=22.12.0",
    "isort>=5.12.0",
    "pydocstyle>=6.3.0",
    "toml>=0.10.2",
    "ruff>=0.0.238",
]

[project.urls]
homepage = "https://momoa.readthedocs.io"
repository = "https://github.com/berislavlopac/momoa"

[tool.black]
line-length = 96
target-version = ['py38', 'py39', 'py310', 'py311']
verbose = false
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 96
filter_files = true
force_alphabetical_sort_within_sections = true

[tool.coverage.run]
source = [ "momoa/", ]
omit = [ "*/tests/*", "src/tests/*", ]

[tool.coverage.report]
skip_covered = false
show_missing = true
fail_under = 90
exclude_lines = [ "pragma: no cover", "@abstract",]

[tool.mypy]
mypy_path = "momoa/"
ignore_missing_imports = true

[tool.pydocstyle]
add-ignore = "D105, D107, D212, D401"
convention = "google"
match-dir = "(?!tests).*"

[tool.ruff]
line-length = 96
target-version = "py38"
format = "grouped"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
